src/microservices/customer-mgmt-service/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
requests==2.26.0
pydantic==1.10.0
pymongo==4.5.0
kafka-python==2.0.2